Implement StyledComponents for rest of site; rename files and folders
