--- src/App.css
@@ -1,13 +1,0 @@
-.main-body {
-    display: flex;
-}
-
-.body-content {
-    width: 100%;
-    padding-top: 25px;
-    margin: 50px;
-}
-
-#di_model {
-    width: 100%;
-}
