+++ src/App.jsx
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import './App.css';
+import { MainBody } from './styles/App';
 import Footer from './components/Footer'
 import {BrowserRouter as Router, Route} from "react-router-dom";
 import HeaderBar from './components/HeaderBar';
@@ -14,24 +14,22 @@ import { faHome, faLightbulb, faHandshake, faAddressCard } from '@fortawesome/fr
 
 library.add(faLinkedin, faTwitter, faInstagram, faHome, faLightbulb, faHandshake, faAddressCard);
 
-class App extends Component {
+export default class App extends Component {
   render() {
       return (
           <Router>
               <div className="App">
                   <HeaderBar />
-                  <div className="main-body">
+                  <MainBody>
                     <LeftSidebar />
                     <Route exact path="/" component={Home}/>
                     <Route path="/blog" component={Blog}/>
                     <Route path="/services" component={Services}/>
                     <Route path="/contact" component={Contact}/>
-                  </div>
+                  </MainBody>
                   <Footer />
               </div>
           </Router>
       );
   }
-}
-
-export default App;
+};
