+++ src/components/styles/BloggerPosts.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const BlogPostContainer = styled.div`
+  margin: 20px 0;
+`;
