+++ src/components/styles/LoaderSpinner.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const LoadingContainer = styled.div`
+  padding: 20px;
+  display: flex;
+  justify-content: center;
+`;
