--- src/pages/Blog.css
@@ -1,9 +1,0 @@
-.LoadingContainer {
-    padding: 20px;
-    display: flex;
-    justify-content: center;
-}
-
-.BlogPostContainer {
-    margin: 20px 0;
-}
