+++ src/pages/Blog.jsx
@@ -1,17 +1,17 @@
 import React, { Component } from 'react';
 import BloggerPosts from '../components/BloggerPosts';
-import './Blog.css';
+import { BodyContent } from './styles/Pages';
 
 export default class Blog extends Component {
   render() {
     return (
-      <div className="body-content">
+      <BodyContent>
         <h1>
             Blog
         </h1>
         <hr/>
         <BloggerPosts />
-      </div>
+      </BodyContent>
     )
   }
 }
