+++ src/pages/Contact.jsx
@@ -1,13 +1,14 @@
 import React, { Component } from 'react';
+import { BodyContent } from './styles/Pages';
 
 export default class Contact extends Component {
     render() {
         return (
-            <div className="body-content">
+            <BodyContent>
                 <h1>
                     Contact
                 </h1>
-            </div>
+            </BodyContent>
         )
     }
 }
