+++ src/pages/Home.jsx
@@ -1,15 +1,16 @@
 import React, {Component} from 'react';
 import {generateParagraphs} from "../helpers";
 import * as Content from "../content/StaticContent";
+import { BodyContent } from './styles/Pages';
 
 export default class Home extends Component {
   render() {
 
     return (
-      <div className="body-content">
+      <BodyContent>
         <h1>About Me</h1>
         {generateParagraphs(Content.Welcome)}
-      </div>
+      </BodyContent>
     )
 
   }
