+++ src/pages/styles/Pages.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const BodyContent = styled.div`
+  width: 100%;
+  padding-top: 25px;
+  margin: 50px;
+`;
