+++ src/styles/App.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const MainBody = styled.div`
+  display: flex;
+`;
